+++ WordPressService/index.js
@@ -1,6 +1,15 @@
 const fetch = require('node-fetch');
 const { fetchJSON } = require('./fetchJSON');
-const { gitAuthheader, gitDefaultOptions, gitHubMessage, gitPutOptions } = require('./gitHub');
+const {
+    gitAuthheader,
+    gitDefaultOptions,
+    gitHubMessage,
+    gitPutOptions,
+    committer,
+    branchDelete,
+    branchGetHead,
+    githubApiUrl
+} = require('./gitHub');
 const { JSDOM } = require("jsdom");
 const sha1 = require('sha1');
 const fs = require('fs');
@@ -19,19 +28,11 @@ const shaupdate = (file, wp_sha, github_sha) => {
 
 let pinghistory = []; //Used to log updates
 
-const committer = {
-    'name': 'WordPressService',
-    'email': 'data@alpha.ca.gov'
-};
-
-//const masterbranch='synctest3', stagingbranch='synctest3_staging', postTranslationUpdates = false, branchprefix = 'synctest3_deploy_';
-const masterbranch='master', stagingbranch='staging', postTranslationUpdates = true, branchprefix = 'wpservice_deploy_';
+const masterbranch='synctest3', stagingbranch='synctest3_staging', postTranslationUpdates = false, branchprefix = 'synctest3_deploy_';
+//const masterbranch='master', stagingbranch='staging', postTranslationUpdates = true, branchprefix = 'wpservice_deploy_';
 const autoApproveTranslationPrs = true;
 const mergetargets = [masterbranch,stagingbranch];
 const appName = 'WordPressService';
-const githubUser = 'cagov';
-const githubRepo = 'covid19';
-const githubApiUrl = `https://api.github.com/repos/${githubUser}/${githubRepo}/`;
 const githubTranslationPingsPath = `pages/translations/pings`;
 const githubTranslationContentPath = `pages/translations/content`;
 const githubTranslationFlatPath = `pages/translated-posts`;
@@ -106,12 +107,6 @@ const translationUpdateAddPost = (Post, download_path) => {
         translationUpdatePayload.push(translationRow);
     }
 }
-
-const branchGetHeadUrl = branch => `${githubApiUrl}git/refs/heads/${branch}`;
-
-//Return a branch head record
-const branchGetHead = async branch =>
-    fetchJSON(branchGetHeadUrl(branch),gitDefaultOptions());
 
 //create a branch for this update
 const branchCreate = async (filename,mergetarget) => {
@@ -230,22 +225,6 @@ const branchMerge = async (branch, mergetarget, bPrMode, PrTitle, PrLabels) => {
 
             await branchDelete(branch);
         }
-    }
-}
-
-const branchDelete = async branch => {
-    //delete
-    //https://developer.github.com/v3/git/refs/#delete-a-reference
-    const deleteBody = {
-        method: 'DELETE',
-        headers: gitAuthheader()
-    };
-    const branchDeleteResult = await fetch(branchGetHeadUrl(branch), deleteBody);
-
-    if(branchDeleteResult.status===204) {
-        console.log(`BRANCH DELETE Success: ${branch}`);
-    } else {
-        console.log(`BRANCH DELETE N/A: ${branch}`);
     }
 }
